backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.5
celery>=5.3.0
redis>=5.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
reportlab>=4.0.0
